blog index with sanity data
